Separate tabs to respective files
